EP6

+++ funcao.c
@@ -1,5 +1,5 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-float triplo(float  numero){return(numero*3);}
+double triplo(double  numero){return(numero*3.0);}
